priestchat.com advertises that it facilitates chats with 'real catholic priests', but actually its all chatbots (and should you bother reading the ToS, it clearly says so).
It shows you a series of pictures of 'priests' who are 'available to chat', even offers activity information and different language skills to make them look more real. One of those actually is tagged 'AI', which implies the others are not. The bots themselves try to disguse their nature, and if asked, avoid revealing their true nature.

This is in particular troubling because the homepage promises that - catholic - confessions and absolutions are possible by their 'priests', which is clearly false.